Сделал отображение
