+++ index.html
@@ -128,7 +128,7 @@
       </section>
 
       <!-- Здесь будут изображения других пользователей -->
-
+        
     </section>
 
     <!-- Полноэкранный показ изображения -->
@@ -229,6 +229,6 @@
   <template id="messages">
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
-  <script type="module" src="main.js"></script>
+  <script type="module" src="js/main.js"></script>
 </body>
 </html>
