+++ js/util.js
@@ -14,3 +14,4 @@ const getRandomNumbers = function(min, max)
 {
   return Math.floor(Math.random() * (max - min + 1)) + min;
 };
+export {getRandomNumber, getRandomNumbers, checkLength};
